Good working model of login and registration. To add a splash screen later and improve UX
