+++ app/src/main/java/my/project/nostalgia/activities/RegisterActivity.java
@@ -1,6 +1,7 @@
 package my.project.nostalgia.activities;
 import android.app.ProgressDialog;
 import android.content.Intent;
+import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.text.TextUtils;
 import android.util.Patterns;
@@ -13,6 +14,7 @@ import android.widget.Toast;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.preference.PreferenceManager;
 
 import com.google.android.gms.tasks.OnCanceledListener;
 import com.google.android.gms.tasks.OnCompleteListener;
@@ -22,6 +24,10 @@ import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
 
 import my.project.nostalgia.R;
+import my.project.nostalgia.supplementary.memoryEvents;
+
+import static my.project.nostalgia.activities.LoginActivity.APPLICABLE_EVENTS;
+import static my.project.nostalgia.activities.LoginActivity.SEND_USERNAME;
 
 public class RegisterActivity extends AppCompatActivity {
 
@@ -48,7 +54,6 @@ public class RegisterActivity extends AppCompatActivity {
                 Register();
             }
         });
-
     }
 
     private void Register() {
@@ -68,11 +73,13 @@ public class RegisterActivity extends AppCompatActivity {
         if(TextUtils.isEmpty(password)){
             mPassword.setError("Enter your password");return;}
         else if(password.length() < 6){
-            mPassword.setError("Password too short. Atleast let it be 6 characters");return;}
+            mPassword.setError("Password too short/weak");return;}
         if(TextUtils.isEmpty(password2)){
             mPassword.setError("Confirm your password");return;}
         else if(!password.equals(password2)){
             mPassword2.setError("Different Password");return;}
+
+        setGeneralInfo(mName.getText().toString(),new memoryEvents(this).getJoinedEvents());
 
         mProgressDialog.setMessage("Please wait...");
         mProgressDialog.show();
@@ -80,28 +87,26 @@ public class RegisterActivity extends AppCompatActivity {
 
         mAuth.createUserWithEmailAndPassword(email,password)
                 .addOnCompleteListener(this, new OnCompleteListener<AuthResult>() {
-                    @Override
-                    public void onComplete(@NonNull Task<AuthResult> task) {
-                        if(task.isSuccessful()) {
-                            Toast.makeText(RegisterActivity.this, "Successfully registered."
-                                    , Toast.LENGTH_SHORT).show();
-                            startActivity(new Intent(RegisterActivity.this,MemoryListActivity.class));
-                            finish();
-                        }
-                        else{
-                            Toast.makeText(RegisterActivity.this, "Registration Failed!"+task.getException().getMessage()
-                                    , Toast.LENGTH_SHORT).show();
-                            mProgressDialog.dismiss();
-                        }
-                    }
-                });
+            @Override
+            public void onComplete(@NonNull Task<AuthResult> task) {
+                if(task.isSuccessful()) {
+                    Toast.makeText(RegisterActivity.this, "Successfully registered."
+                            , Toast.LENGTH_SHORT).show();
+                    startActivity(new Intent(RegisterActivity.this,MemoryListActivity.class));
+                    finish();
+                }
+                else{
+                    Toast.makeText(RegisterActivity.this, "Registration Failed!"
+                            , Toast.LENGTH_SHORT).show();
+                    mProgressDialog.dismiss();
+                }
+            }
+        });
     }
-
-    @Override
-    protected void onStart() {
-        super.onStart();
-        FirebaseUser currentUser = mAuth.getCurrentUser();
-        if(currentUser!=null)
-            Toast.makeText(this,"Already Signed in",Toast.LENGTH_SHORT).show();
+    private void setGeneralInfo(String userName, String combinedEvents) {
+        SharedPreferences.Editor editor = PreferenceManager.getDefaultSharedPreferences(this).edit();
+        editor.putString(SEND_USERNAME, userName);
+        editor.putString(APPLICABLE_EVENTS, combinedEvents);
+        editor.apply();
     }
 }
